refactor: move limb composition to package
